Move animation logic and event handling to hook
