+++ packages/harpguru-core/src/components/menu-open-button/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useTapAnimation } from './use-tap-animation'
